reeduc/requirements.txt
Django==6.0.1
python-dotenv==1.0.1
psycopg[binary]==3.2.1
